+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## 0.1.3 (2021-12-03)
+
+* Ability to reach all endpoints from one `OpenseaAPI` object
+* API key support (Opensea requires it from now on)
+
 ## 0.1.0 (2021-11-07)
 
 * First release on PyPI.